+external_time string column
get and get_nodefault and append now take an optional table_type argument

added get_indexed to allow the user to use cassandra secondary indices

selectTimeInterval keeps giving me rows that are outside the requested interval, or even for a different item! As a stopgap, i added manual filtering. For a large database, this will not scale (in which case, why use Cassandra at all?), so ultimately a better solution must be found. My current best guess is that we need to leave pycassa, which deprecates itself on its webpage, and use the CQL-based http://datastax.github.io/python-driver/ instead; hopefully the CQL-based driver handles ordered range scans over composite keys differently from, and more successfully than pycassa. But it's possible that i am trying to do something that Cassandra just doesn't support at all.

as a test, i tried making a selectTimeInterval2, which uses secondary indices to filter the times instead of a range scan. It appears to loads all rows for the given item into memory at once and then scan the time in memory. This is not scalable. I don't recommend using it.